CH: Moving product header to new file
